Refactoring thirdparty libs into a separate module.
