+++ .flake8
@@ -3,6 +3,6 @@ ignore = W503,E402,E731
 exclude =
     .git, __pycache__, build, dist, .eggs, .github, .local, docs/,
     Samples, azure_functions_worker/protos/,
-    azure_functions_worker/typing_inspect.py,
+    azure_functions_worker/_thirdparty/typing_inspect.py,
     tests/unittests/test_typing_inspect.py,
     .venv*, .env*, .vscode, venv
