+++ setup.cfg
@@ -13,11 +13,11 @@ warn_return_any = True
 disallow_subclassing_any = False
 ignore_missing_imports = True
 
-[mypy-azure_functions_worker.aio_compat]
+[mypy-azure_functions_worker._thirdparty.aio_compat]
 ignore_errors = True
 
 [mypy-azure_functions_worker.protos.*]
 ignore_errors = True
 
-[mypy-azure_functions_worker.typing_inspect]
+[mypy-azure_functions_worker._thirdparty.typing_inspect]
 ignore_errors = True
