+++ tests/unittests/test_typing_inspect.py
@@ -2,12 +2,11 @@
 # Author: Ivan Levkivskyi
 # License: MIT
 
-from azure_functions_worker.typing_inspect import (
+from azure_functions_worker._thirdparty.typing_inspect import (
     is_generic_type, is_callable_type, is_tuple_type, is_union_type,
     is_typevar, is_classvar, get_origin, get_parameters, get_last_args, get_args,
-    get_generic_type, get_generic_bases, get_last_origin, typed_dict_keys,
-)
-from unittest import TestCase, main, skipIf, skipUnless
+    get_generic_type, get_generic_bases, get_last_origin, )
+from unittest import TestCase, main, skipIf
 from typing import (
     Union, ClassVar, Callable, Optional, TypeVar, Sequence, Mapping,
     MutableMapping, Iterable, Generic, List, Any, Dict, Tuple, NamedTuple,
